#96: split `/info` back into individual commands
